Added CVV format
The Card Verification Value (CVV) is an extra code printed on your debit or credit card.
CVV for Visa, MasterCard and Diners is the final three digits of the number printed on the signature strip on the back of your card.
CVV for American Express appears as a separate 4-digit code printed on the front of your card.

This PR extends the format class to include a custom validator for the CVV number.

To use, just specify a property in this way:
```
{
  [...],
  "properties": {
    "cvv": {
      "type": "string",
      "format": "cvv"
    }
  },
  [...]
}
```